requirements.txt
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.16
anndata==0.10.7
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.6
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
ax-platform==0.4.0
Babel==2.14.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.3
bel-resources==0.0.3
biopython==1.78
bioregistry==0.11.9
biothings-client==0.3.1
bleach==6.1.0
blinker==1.7.0
botocore==1.34.69
botorch==0.11.0
cachetools==5.3.3
catboost==1.2.5
cellxgene-census==1.10.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chemfunc==1.0.4
chemprop==1.6.1
click==8.1.7
click-plugins==1.1.1
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.0
curies==0.7.9
cycler==0.12.1
dataclasses==0.6
datamol==0.12.5
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
DeepPurpose==0.1.5
defusedxml==0.7.1
descriptastorus==2.6.1
dgl==2.1.0
dgllife==0.3.2
docstring_parser==0.16
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup @ file:///croot/exceptiongroup_1668714342571/work
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.0
fonttools==4.50.0
formulaic==1.0.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
fuzzywuzzy==0.18.0
gcsfs==2024.6.0
gget==0.28.4
google-api-core==2.19.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gpytorch==1.11
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.63.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
humanize==4.9.0
hyperopt==0.2.7
idna==3.6
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
interface-meta==1.3.0
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///croot/ipython_1694181358621/work
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jaxtyping==0.2.28
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lifelines==0.28.0
lightgbm==4.3.0
linear-operator==0.5.1
llvmlite==0.42.0
loguru==0.7.2
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==3.0.2
mock==5.1.0
molfeat==0.10.0
mordredcommunity==2.0.5
more-click==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multipledispatch==1.0.0
multisplitby==0.0.1
mygene==3.2.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.29
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==3.2.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.7.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.0
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==0.5.6
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow==10.2.0
platformdirs==4.2.0
plotly==5.22.0
pmapper==1.1.1
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
proto-plus==1.24.0
protobuf==5.26.0
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
psycopg2-binary==2.9.9
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py4j==0.10.9.7
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybel==0.15.5
PyBioMed==1.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments @ file:///croot/pygments_1684279966437/work
pynndescent==0.5.12
pyparsing==3.1.2
pyre-extensions==0.0.30
pyro-api==0.1.2
pyro-ppl==1.9.0
pystow==0.5.4
PyTDC==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyTrie==0.4.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
ratelimit==2.2.1
rdkit==2023.3.3
rdkit-pypi==2022.9.5
referencing==0.34.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3fs==2024.6.0
scanpy==1.9.6
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.0
selfies==2.1.1
Send2Trash==1.8.2
session-info==1.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
snowballstemmer==2.2.0
somacore==1.0.7
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.31
stack-data==0.6.3
statsmodels==0.14.2
stdlib-list==0.10.0
subword-nmt==0.3.8
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.4.0
tiledb==0.25.1
tiledbsoma==1.7.2
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.1
torch_geometric==2.5.3
torchdata==0.7.1
torchsummary==1.5.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
triton==2.2.0
typed-argument-parser==1.9.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2024.2.0
xgboost==2.0.3
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1